postgres_conn/pyproject.toml
[project]
name = "postgres_conn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "psycopg2>=2.9.11",
    "python-dotenv>=1.2.1",
]
